Plotting code moved to new package.
