+++ OSRS_hiscores.py
@@ -9,32 +9,15 @@ Created on 14 April 2022
 
 import argparse
 import sys
-import pandas as pd
-from plotly.subplots import make_subplots
-from plotly.offline import plot
-import plotly.graph_objects as go
-import seaborn as sns
 import matplotlib.pyplot as plt
-import matplotlib.patches
 
 # Own packages
 from config.config import read_config, print_config
 from player.player import Player
+import plotting
 
 # Plotting options
 plt.style.use('ggplot')
-
-
-def RotateTickLabels(fig):
-    """
-    Rotates all x-axis tick labels in seaborn facetgrid
-
-    Parameters
-    ----------
-    fig : seaborn facetgrid
-    """
-    for axes in fig.axes.flat:
-        _ = axes.set_xticklabels(axes.get_xticklabels(), rotation=90)
 
 
 if __name__ == '__main__':
@@ -100,63 +83,12 @@ if __name__ == '__main__':
 
     # Create interactive plots of overall level/XP and total boss kill count
     # Note that this plot will open in the default browser
-    overall_plot = make_subplots(rows=2, cols=2,
-                                 vertical_spacing=0.2,
-                                 specs=[[{}, {}],
-                                        [{"colspan": 2}, None]])
-    overall_plot.append_trace(go.Scatter(x=skills[(skills['Skill'] ==
-                                                   'Overall')]['Date'],
-                                         y=skills[(skills['Skill'] ==
-                                                   'Overall')]['Level'],
-                                         name='Level', mode='lines'),
-                              row=1, col=1)
-    overall_plot.append_trace(go.Scatter(x=skills[(skills['Skill'] ==
-                                                   'Overall')]['Date'],
-                                         y=skills[(skills['Skill'] ==
-                                                   'Overall')]['XP'],
-                                         name='XP', mode='lines'),
-                              row=1, col=2)
-    overall_plot.append_trace(go.Scatter(x=total_killcount['Date'],
-                                         y=total_killcount['Kill count'],
-                                         name='Kill count', mode='lines'),
-                              row=2, col=1)
-    overall_plot.update_xaxes(title_text='Date', row=1, col=1)
-    overall_plot.update_xaxes(title_text='Date', row=1, col=2)
-    overall_plot.update_xaxes(title_text='Date', row=2, col=1)
-    overall_plot.update_yaxes(title_text='Total level', row=1, col=1)
-    overall_plot.update_yaxes(title_text='Total XP', row=1, col=2)
-    overall_plot.update_yaxes(title_text='Total boss kill count', row=2, col=1)
-    overall_plot.update_layout(title_text='Overall stats')
-    plot(overall_plot, auto_open=True, filename='overall_stats.html')
+    plotting.summary.create_summary_plots(player)
 
     # Plot skills on facetgrid (one skill per plot)
-    skills_plot = sns.FacetGrid(data=skills, col='Skill', col_wrap=4,
-                                sharey=False, height=3.5, aspect=1.5)
-    skills_plot.map(sns.lineplot, 'Date', 'XP')
-    # Add plot of level to XP plot using a 2nd y-axis
-    for ax, (_, subdata) in zip(skills_plot.axes,
-                                skills.groupby('Skill', sort=False)):
-        ax2 = ax.twinx()
-        subdata.plot(x='Date', y='Level', ax=ax2, legend=False, color='b')
-        ax.set_ylabel('XP')
-        ax2.set_ylabel('Level')
-
-        # Create legend (colours obtained via plt.gca().lines[-1].get_color())
-        name_to_color = {
-            'XP': (0.8862745098039215, 0.2901960784313726, 0.2),
-            'Level': 'b'
-        }
-        patches = [matplotlib.patches.Patch(
-            color=v, label=k) for k, v in name_to_color.items()]
-        plt.legend(handles=patches, loc='upper left')
-    RotateTickLabels(skills_plot)
-    plt.tight_layout()
+    plotting.grid_plots.plot_skills(player)
 
     # Plot killcount on facetgrid (one boss per plot)
-    killcount_plot = sns.FacetGrid(data=killcount, col='Boss', col_wrap=6,
-                                   sharey=False)
-    killcount_plot.map(sns.lineplot, 'Date', 'Kill count')
-    RotateTickLabels(killcount_plot)
-    plt.tight_layout()
+    plotting.grid_plots.plot_killcount(player)
 
     plt.show()
