+++ plotting/__init__.py
@@ -1,0 +1,3 @@
+# plotting/__init__.py
+import plotting.summary
+import plotting.grid_plots
